refactor: Organize codes
