+++ src/components/salespost/address/styled.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const SelectWrapper = styled.div``;
+
+export const SelectBox = styled.div`
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+	margin: 20px;
+	margin-left: 35px;
+`;
