+++ src/components/salespost/atoms.js
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const imageFiles = atom({
+	key: 'imgFiles',
+	default: [],
+});
