+++ src/index.js
@@ -1,14 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
+import { RecoilRoot } from 'recoil';
 import App from './App';
 
 const root = document.getElementById('root');
 
 ReactDOM.createRoot(root).render(
 	<React.StrictMode>
-		<BrowserRouter>
-			<App />
-		</BrowserRouter>
+		<RecoilRoot>
+			<BrowserRouter>
+				<App />
+			</BrowserRouter>
+		</RecoilRoot>
 	</React.StrictMode>,
 );
